requirements.txt
# Nexus Distributor Support — production dependencies
Django>=4.2,<5.0
djangorestframework>=3.14,<3.16
django-filter>=24.2
django-cors-headers>=4.3
django-environ>=0.11
psycopg[binary]>=3.1,<3.3
celery[redis]>=5.3,<5.5
redis>=5.0,<6
gunicorn>=22.0
whitenoise[brotli]>=6.6
Pillow>=10.3
cryptography>=42.0
pyjwt>=2.8
requests>=2.31
django-ratelimit>=4.1
